add an argment

+++ mvgraph.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 
-mv ./import_graph.dot ../data
+if [ -f "./import_graph.dot" ]; then
+    mv ./import_graph.dot ../data
+fi
+
+cp -r ./mathlib4 ../data
+
+tail -f /dev/null
